+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- char count feature
+- manupaper count feature
+### Changed
+- build core tools
 
 ## [0.0.1] - 2021-03-01
 ### Added